refactor: extract prompt building, response cleaning, and history export from FFAI into src/core modules
Extract PromptBuilder, HistoryExporter, prompt_utils, and response_utils
into dedicated src/core/ modules. Update all imports across clients,
orchestrator, and agent to use new paths. Clean up verbose logging in
ordered history and add type hints to conversation/permanent history.
