+++ src/Clients/FFMistralSmall.py
@@ -12,7 +12,7 @@ from typing import Any
 from dotenv import load_dotenv
 from mistralai import Mistral
 
-from ..FFAIClientBase import FFAIClientBase
+from ..core.client_base import FFAIClientBase
 from ..retry_utils import (
     create_rate_limit_error,
     get_retry_decorator,
