+++ src/core/history/__init__.py
@@ -1,0 +1,15 @@
+# Copyright (c) 2025 Antonio Quinonez / Far Finer LLC
+# SPDX-License-Identifier: MIT
+# Contact: antquinonez@farfiner.com
+
+"""History tracking strategies for FFAI interactions."""
+
+from .conversation import ConversationHistory
+from .ordered import OrderedPromptHistory
+from .permanent import PermanentHistory
+
+__all__ = [
+    "ConversationHistory",
+    "OrderedPromptHistory",
+    "PermanentHistory",
+]
